replaced models.ScoreDenoiser for optim.ScorePrior
