+++ deepinv/optim/__init__.py
@@ -1,3 +1,4 @@
 from .data_fidelity import DataFidelity, L2, L1, IndicatorL2, PoissonLikelihood
 from .optimizers import BaseOptim, optim_builder
 from .fixed_point import FixedPoint, AndersonAcceleration
+from .prior import Prior, ScorePrior, Tikhonov, PnP, RED
