+++ deepinv/tests/test_sampling.py
@@ -4,8 +4,6 @@ import torch.nn
 import deepinv as dinv
 from deepinv.optim.data_fidelity import L2
 from deepinv.sampling import ULA, SKRock
-from deepinv.tests.dummy_datasets.datasets import DummyCircles
-from torch.utils.data import DataLoader
 import numpy as np
