+++ examples/sampling/demo_custom_kernel.py
@@ -176,7 +176,7 @@ class PreconULA(dinv.sampling.MonteCarlo):
 
 model_spec = {"name": "median_filter", "args": {}}
 
-prior = dinv.models.ScoreDenoiser(model_spec=model_spec)
+prior = dinv.optim.ScorePrior(denoiser=dinv.models.Denoiser(model_spec=model_spec))
 
 
 # %%
@@ -184,7 +184,6 @@ prior = dinv.models.ScoreDenoiser(model_spec=model_spec)
 # ---------------------------------------------------
 # We create the preconditioned and standard ULA samplers using
 # the same hyperparameters (step size, number of iterations, etc.).
-
 
 step_size = 0.5 * (sigma**2)
 iterations = int(1e2) if torch.cuda.is_available() else 10
